% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{add_label}
\alias{add_label}
\title{Helper function: add label ("SiteName_ParaName_Unit")}
\usage{
add_label(df, col_sitename = "SiteName",
  col_parametername = "ParameterName", col_parameterunit = "ParameterUnit")
}
\arguments{
\item{df}{data frame containing at least a columns "SiteName", "ParameterName",
"ParameterUnit"}

\item{col_sitename}{column in df containing site name (default: "SiteName")}

\item{col_parametername}{column in df containing parameter name (default: "ParameterName")}

\item{col_parameterunit}{column in df containing parameter unit (default: "ParameterUnit")}
}
\value{
returns input data frame with added column "SiteName_ParaName_Unit"
}
\description{
Helper function: add label ("SiteName_ParaName_Unit")
}
