% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_periods.R
\name{get_rawfilespaths_for_month}
\alias{get_rawfilespaths_for_month}
\title{Berlin-Tiefwerder: get rawfilepaths for months}
\usage{
get_rawfilespaths_for_month(monthly_period = get_monthly_periods()[1, ],
  raw_data_dir = system.file("shiny/berlin_t/data/operation", package =
  "aquanes.report"), max_offset_days = 7)
}
\arguments{
\item{monthly_period}{one row of data.frame as retrieved by function
first row of get_monthly_periods(), i.e. year month is (default: '2017-06')}

\item{year_month_end}{end year month (default: current month )}

\item{tz}{(default: 'CET')}
}
\value{
dataframe with monthly periods
}
\description{
Berlin-Tiefwerder: get rawfilepaths for months
}
