% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_datetime.R
\name{group_datetime}
\alias{group_datetime}
\title{Group DateTime by user defined period (year, month, day, hour, minute)}
\usage{
group_datetime(df, by = "day", fun = "median", col_datetime = "DateTime",
  col_datatype = "DataType")
}
\arguments{
\item{df}{a data frame as retrieved by import_data_haridwar()
(default: system.file("shiny/haridwar/data/161101Monitoring_AquaNES_4014_Haridwar_KWB.xlsx",
package = "aquanes.report"))}

\item{by}{an aggregation time step of "year", "month", "day", "hour", "minute"
(default: "day)}

\item{fun}{function to be used for grouping measurement data of column ParameterValue
(default: median)
(default: system.file("shiny/haridwar/.my.cnf", package = "aquanes.report"))}

\item{col_datetime}{column name of datetime column (default: DateTime)}

\item{col_datatype}{column name of data type column (default: DateType)}
}
\value{
returns data frame with data aggregated according to user defined
aggregation time step
}
\description{
Group DateTime by user defined period (year, month, day, hour, minute)
}
