% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_t.R
\name{import_data_berlin_t}
\alias{import_data_berlin_t}
\title{Import data for Berlin Tiefwerder}
\usage{

  import_data_berlin_t(raw_data_dir = system.file("shiny/berlin_t/data/operation",
  package = "aquanes.report"),
  meta_file_path = system.file("shiny/berlin_t/data/parameter_site_metadata.csv",
  package = "aquanes.report"),
  fts_file_path = system.file("shiny/berlin_t/data/siteData_raw_list.fts",
  package = "aquanes.report"))
}
\arguments{
\item{raw_data_dir}{path of directory containing PENTAIR xls files (default:
(default: system.file("shiny/berlin_t/data/operation",
package = "aquanes.report"))))}

\item{meta_file_path}{path to metadata file (default:
system.file("shiny/berlin_t/data/parameter_site_metadata.csv", package =
"aquanes.report")))}

\item{fts_file_path}{path to fts file (default:
system.file("shiny/berlin_t/data/siteData_raw_list.fts", package =
"aquanes.report")))}
}
\value{
list with "df": data.frame with imported operational data (analytics
data to be added as soon as available) and "added_data_points": number of
added data points in case of existing fts file was updated with new operational
data
}
\description{
Import data for Berlin Tiefwerder
}
