% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_haridwar}
\alias{import_data_haridwar}
\title{Imports Haridwar data}
\usage{
import_data_haridwar(analytics_path = system.file(file.path("shiny",
  "haridwar/data/161101Monitoring_AquaNES_4014_Haridwar_KWB.xlsx"), package =
  "aquanes.report"),
  operation_mySQL_conf = system.file("shiny/haridwar/.my.cnf", package =
  "aquanes.report"),
  operation_meta_path = system.file(file.path("shiny/haridwar/data",
  "operation_parameters.csv"), package = "aquanes.report"), debug = TRUE)
}
\arguments{
\item{analytics_path}{Define path of analytics EXCEL spreadsheet to be
imported (default: system.file(file.path("shiny/haridwar/data",
"161101Monitoring_AquaNES_4014_Haridwar_KWB.xlsx"),
package = "aquanes.report"))}

\item{operation_mySQL_conf}{column name pattern for identifying raw data
(default: system.file("shiny/haridwar/.my.cnf", package = "aquanes.report"))}

\item{operation_meta_path}{path to table with meta data for operational
parameters (default: system.file(file.path("shiny/haridwar/data",
"operation_parameters.csv"), package = "aquanes.report"))}

\item{debug}{if TRUE print debug messages (default: TRUE)}
}
\value{
returns data frame with Haridwar raw data (operation & analytics)
}
\description{
Imports Haridwar data
}
