% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_julo.R
\name{import_operation_julo}
\alias{import_operation_julo}
\title{Imports operational data for Julo}
\usage{
import_operation_julo(mysql_conf = file.path(getwd(), ".my.cnf"))
}
\arguments{
\item{mysql_conf}{path to the MySQL configuration file}
}
\value{
returns data frame operational data from MySQL db
}
\description{
Imports operational data for Julo
}
