% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_t.R
\name{read_pentair_data}
\alias{read_pentair_data}
\title{Read PENTAIR operational data}
\usage{
read_pentair_data(raw_data_dir = system.file("shiny/berlin_t/data/operation",
  package = "aquanes.report"),
  meta_file_path = system.file("shiny/berlin_t/data/parameter_site_metadata.csv",
  package = "aquanes.report"))
}
\arguments{
\item{raw_data_dir}{path of directory containing PENTAIR xls files (default:
(default: system.file("shiny/berlin_t/data/operation",
package = "aquanes.report"))))}

\item{meta_file_path}{path to metadata file (default:
system.file("shiny/berlin_t/data/parameter_site_metadata.csv", package =
"aquanes.report")))}
}
\value{
data.frame with imported PENTAIR operational data
}
\description{
Read PENTAIR operational data
}
