% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_s.R
\name{read_wedeco_data}
\alias{read_wedeco_data}
\title{Import WEDECO raw data}
\usage{
read_wedeco_data(raw_data_dir = system.file("shiny/berlin_s/data/operation",
  package = "aquanes.report"),
  meta_file_path = system.file("shiny/berlin_s/data/parameter_site_metadata.csv",
  package = "aquanes.report"))
}
\arguments{
\item{raw_data_dir}{path to raw data directory}

\item{meta_file_path}{path to meta data file}
}
\description{
Import WEDECO raw data
}
