% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shinyapp}
\alias{run_shinyapp}
\title{Runs Shiny app for an AQUANES site}
\usage{
run_shinyapp(siteName = "haridwar", use_live_data = FALSE,
  launch.browser = TRUE, ...)
}
\arguments{
\item{siteName}{site name for shiny app (default: "haridwar")}

\item{use_live_data}{should live data be used (default: FALSE)}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started (default: TRUE)}

\item{...}{further arguments passed to shiny::runApp()}
}
\description{
Runs Shiny app for an AQUANES site
}
