% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shinyapp}
\alias{run_shinyapp}
\title{Runs Shiny App}
\usage{
run_shinyapp(appName = "timeseries", launch.browser = TRUE, ...)
}
\arguments{
\item{appName}{name of shiny app (default: "visualise")}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started (default: TRUE)}

\item{...}{further arguments passed to shiny::runApp()}
}
\description{
Runs Shiny App
}
